refactor: move summary types to cluster_summary.go
Moved all cluster summary related type definitions from cluster_meta.go to
cluster_summary.go for better organization and maintainability.
